% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSingleCellExperiment.R
\name{loadSingleCellExperiment}
\alias{loadSingleCellExperiment}
\title{Load a SingleCellExperiment}
\usage{
loadSingleCellExperiment(exp.info, project, ...)
}
\arguments{
\item{exp.info}{Named list containing the metadata for this experiment.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}. 
By default, this should be a string containing the path to a staging directory.}

\item{...}{Further arguments to pass to \code{\link{loadSummarizedExperiment}}.}
}
\value{
A \linkS4class{SingleCellExperiment} object.
}
\description{
Load a \linkS4class{SingleCellExperiment} object from file, using the metadata generated by the corresponding \code{\link{stageObject}} method.
}
\examples{
# Mocking up an SCE:
mat <- matrix(rpois(10000, 10), ncol=10)
colnames(mat) <- letters[1:10]
rownames(mat) <- sprintf("GENE_\%i", seq_len(nrow(mat)))

se <- SingleCellExperiment(list(counts=mat))
se$stuff <- LETTERS[1:10]
se$blah <- runif(10)
reducedDims(se) <- list(
    PCA=matrix(rnorm(ncol(se)*10), ncol=10),
    TSNE=matrix(rnorm(ncol(se)*2), ncol=2)
)
altExps(se) <- list(spikes=SummarizedExperiment(list(counts=mat[1:2,])))

# Staging it: 
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(se, dir=tmp, "rna-seq") 

# Loading it back into memory:
loadSingleCellExperiment(info, tmp)

}
\author{
Aaron Lun
}
